% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_style_specs.R
\name{combine_style_specs}
\alias{combine_style_specs}
\title{Combine two \code{style_specs} lists}
\usage{
combine_style_specs(s1, s2)
}
\arguments{
\item{s1}{\code{style_spec} list 1.}

\item{s2}{\code{style_spec} list 2.}
}
\value{
List of class "style_spec".
}
\description{
Compare each value of \code{s1} against \code{s2}. If both are populated (not
\code{NULL}), then use \code{s1}'s value. Otherwise, one of them is
\code{NULL} so use the other.
}
