% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_borders}
\alias{HTable_borders}
\title{Add borders to the table}
\usage{
obj$borders()
}
\arguments{
\item{line}{Character vector (length 1) of line type to use for the borders.}

\item{width}{Character vector (length 1) of border width. Usually in the
format, "Npx".}

\item{color}{Character vector (length 1) of border color.}
}
\description{
Add borders around all cells.
}
