% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_styles.R
\name{HTable_cell_add_style}
\alias{HTable_cell_add_style}
\title{Add a style to a cell}
\usage{
obj$cell_add_style(row = NULL, col = NULL, style = NULL,
  include_header = FALSE)
}
\arguments{
\item{row}{Numeric vector of which rows to target.}

\item{col}{Numeric vector of which columns to target.}

\item{style}{Character vector (length 1) of style(s).}

\item{include_header}{Boolean if \code{rows} should include the header as the
1st index. E.g. if \code{FALSE}, \code{row = 1} will target the first row,
but if \code{TRUE}, \code{row = 1} will target the header.}
}
\description{
Add a style to a specific <td> or <th> tag. Changes the
  \code{styles} field.
}
