% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_styles.R
\name{HTable_cell_replace_style}
\alias{HTable_cell_replace_style}
\title{Replace the style of a cell}
\arguments{
\item{row}{Numeric vector of which rows to target.}

\item{col}{Numeric vector of which columns to target.}

\item{style}{Character vector (length 1) of style(s).}

\item{include_header}{Boolean if \code{rows} should include the header as the
1st index. E.g. if \code{FALSE}, \code{row = 1} will target the first row,
but if \code{TRUE}, \code{row = 1} will target the header.}
}
\description{
Replace the style of a specific <td> or <th> tag. Changes the \code{styles}
field.
}
