% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_clear_styles}
\alias{HTable_clear_styles}
\title{Clear the style of every <th> and <td> tag}
\description{
Clear the style of every entry in the \code{styles} field. This is a
character matrix of styles that represents the styles applied to each <th>
and <td> tag.
}
