% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_styles.R
\name{HTable_col_add_style}
\alias{HTable_col_add_style}
\title{Add a style to a column}
\usage{
obj$col_add_style(col = NULL, style = NULL, include_header = FALSE)
}
\arguments{
\item{col}{Numeric or character vector of which columns to target.}

\item{style}{Character vector (length 1) of style(s).}

\item{include_header}{Boolean if the <th> tags in the columns should also be
styled.}
}
\description{
Add a style to each <td> and, optionally, <th> tag in a column.
  Changes the \code{styles} field.
}
