% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_styles.R
\name{HTable_col_bg_color}
\alias{HTable_col_bg_color}
\title{Add a background color style to a column}
\usage{
obj$col_bg_color(col = NULL, color = NULL, include_header = FALSE)
}
\arguments{
\item{col}{Numeric or character vector of which columns to target.}

\item{color}{Character vector (length 1) of an HTML color name, hex color
code, or rgb color of the form rgb(x, y, z).}

\item{include_header}{Boolean if the <th> tags in the columns should also be
styled.}
}
\description{
Add a background color style to each <td> and, optionally, <th>
  tag in a column. Changes the \code{styles} field.
}
