% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_styles.R
\name{HTable_col_dollar_fmt}
\alias{HTable_col_dollar_fmt}
\title{Edit a column to have a dollar format}
\usage{
obj$col_dollar_fmt(col = NULL)
}
\arguments{
\item{col}{Numeric or character vector of which columns to target.}
}
\description{
Change numeric column(s) to have a dollar formatting. Changes
  the \code{contents} field.
}
