% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_styles.R
\name{HTable_col_pct_fmt}
\alias{HTable_col_pct_fmt}
\title{Edit a column to have a percent format}
\arguments{
\item{col}{Numeric vector of which columns to target.}
}
\description{
Change numeric column(s) to have a percent formatting. Changes the
\code{contents} field.
}
