% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_styles.R
\name{HTable_div_max_height}
\alias{HTable_div_max_height}
\title{Set the max height of the <div> tag}
\usage{
obj$div_max_height()
}
\arguments{
\item{height}{Character vector (length 1) of the max height to use. Usually
of the format "Npx" or "N\%".}
}
\description{
Set the max height of the <div> tag. Changes the
  \code{div_style} field.
}
