% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_styles.R
\name{HTable_div_max_width}
\alias{HTable_div_max_width}
\title{Set the max width of the <div> tag}
\usage{
obj$div_max_width()
}
\arguments{
\item{width}{Character vector (length 1) of the max width to use. Usually of
the format "Npx" or "N\%".}
}
\description{
Set the max width of the <div> tag. Changes the \code{div_style}
  field.
}
