% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_styles.R
\name{HTable_div_style_add}
\alias{HTable_div_style_add}
\title{Add a style to the <div> tag}
\usage{
obj$div_style_add(style = NULL)
}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Add to the existing styles located in the \code{div_style}
  field.
}
