% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_styles.R
\name{HTable_div_style_replace}
\alias{HTable_div_style_replace}
\title{Replace the <div> tag style}
\usage{
obj$div_style_replace(style = NULL)
}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Overwrite the \code{div_style} field.
}
