% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_styles.R
\name{HTable_div_x_scroll}
\alias{HTable_div_x_scroll}
\title{Add an auto scroll bar to the <div> tag in the x direction}
\usage{
obj$div_x_scroll()
}
\description{
Add an auto scroll bar to the <div> tag in the x direction.
  Changes the \code{div_style} field.
}
