% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_font_family}
\alias{HTable_font_family}
\title{Change the font family}
\usage{
obj$font_family()
}
\arguments{
\item{size}{Character vector (length 1) of font family.}
}
\description{
Change the font family of all cells.
}
