% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_font_size}
\alias{HTable_font_size}
\title{Change the font size}
\usage{
obj$font_size()
}
\arguments{
\item{size}{Character vector (length 1) of font size. Usually in the format,
"Npt".}
}
\description{
Change the font size of all cells.
}
