% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_header_add_style}
\alias{HTable_header_add_style}
\title{Add a style to the header}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Add a style to each <th> tag. Changes the \code{styles} field.
}
