% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_header_bg_color}
\alias{HTable_header_bg_color}
\title{Add a background color style to the header}
\arguments{
\item{color}{Character vector (length 1) of an HTML color name, hex color
code, or rgb color of the form rgb(x, y, z).}
}
\description{
Add a background color style to each <th> tag. Changes the \code{styles}
field.
}
