% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_header_replace_style}
\alias{HTable_header_replace_style}
\title{Replace the style of the header}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Replace the style of each <th> tag. Changes the \code{styles} field.
}
