% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_padding}
\alias{HTable_padding}
\title{Add padding to the table}
\usage{
obj$padding()
}
\arguments{
\item{padding}{Character vector (length 1) of padding distance. Usually in
the format, "Npx".}
}
\description{
Add padding to all cells.
}
