% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htable.R
\name{HTable_render}
\alias{HTable_render}
\title{Generate the HTML of the HTable object}
\usage{
obj$render("html")
}
\arguments{
\item{type}{Render either an HTML or Markdown table.}
}
\value{
A length 1 character vector.
}
\description{
Generate the HTML of the HTable object
}
