% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_reset_contents}
\alias{HTable_reset_contents}
\title{Reset the contents of every <th> and <td> tag}
\description{
Reset the contents of every entry in the \code{contents} field. This is a
character matrix of contents that represents the contents of each <th> and
<td> tag.
}
