% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htable.R
\name{HTable_Rmd}
\alias{HTable_Rmd}
\title{Put an HTable into an RMarkdown Document}
\usage{
obj$Rmd()
}
\arguments{
\item{type}{Include either an HTML or Markdown table.}
}
\description{
For use in an R Markdown document inside of a chunk with the
  option \code{results='asis'}.
}
