% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_row_alt_bg_color}
\alias{HTable_row_alt_bg_color}
\title{Add an alternating background color style to all rows}
\arguments{
\item{color1}{Character vector (length 1) of an HTML color name, hex color
code, or rgb color of the form rgb(x, y, z).}

\item{color2}{Character vector (length 1) of an HTML color name, hex color
code, or rgb color of the form rgb(x, y, z).}

\item{include_header}{Boolean if the header should also be given a background
color.}
}
\description{
Add an alternating background color style to each <td> and, optionally, <th>
tag. The odd rows get the background color \code{color1} and the evens get
the background color \code{color2}. Changes the \code{styles} field.
}
