% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_row_clear_style}
\alias{HTable_row_clear_style}
\title{Clear the style in a row}
\usage{
obj$row_clear_style(row = NULL, include_header = FALSE)
}
\arguments{
\item{row}{Numeric vector of which rows to target.}

\item{style}{Character vector (length 1) of a style to use.}

\item{include_header}{Boolean if \code{rows} should include the header as the
1st index. E.g. if \code{FALSE}, \code{row = 1} will target the first row,
but if \code{TRUE}, \code{row = 1} will target the header.}
}
\description{
Clear the style of each <td> and, optionally, <th> tag in a row.
  Changes the \code{styles} field.
}
