% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_row_color}
\alias{HTable_row_color}
\title{Add a text color style to a row}
\arguments{
\item{row}{Numeric vector of which rows to target.}

\item{color}{Character vector (length 1) of an HTML color name, hex color
code, or rgb color of the form rgb(x, y, z).}

\item{include_header}{Boolean if \code{rows} should include the header as the
1st index. E.g. if \code{FALSE}, \code{row = 1} will target the first row,
but if \code{TRUE}, \code{row = 1} will target the header.}
}
\description{
Add a text style to each <td> and, optionally, <th> tag in a row. Changes the
\code{styles} field.
}
