% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_row_color_scale}
\alias{HTable_row_color_scale}
\title{Apply a color scale gradient to rows}
\usage{
obj$row_color_scale(row = NULL, color = c("#63BE7B", "#FFEB84",
  "#F8696B"), exclude_cols = NULL, na.rm = TRUE, all = FALSE)
}
\arguments{
\item{row}{Numeric vector of which rows to target. \code{row = 1} corresponds
to the first row after the header row - the header row cannot be targeted
with color scales.}

\item{color}{Character vector of HTML color names, hex color codes, or rgb
colors of the form rgb(x, y, z). Used to create a color palette.}

\item{exclude_cols}{Numeric vector of which columns to exclude from
calculation and coloring.}

\item{all}{Boolean if the color scaling should be scaled across all of the
values in the \code{row} rows.}
}
\description{
Change numeric values in row(s) to have a background color
  gradient based on the range of values in the row.
}
