% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_row_italic}
\alias{HTable_row_italic}
\title{Add an italic style to a row}
\arguments{
\item{row}{Numeric vector of which rows to target.}

\item{include_header}{Boolean if \code{rows} should include the header as the
1st index. E.g. if \code{FALSE}, \code{row = 1} will target the first row,
but if \code{TRUE}, \code{row = 1} will target the header.}
}
\description{
Add an italic style to each <td> and, optionally, <th> tag in a row. Changes
the \code{styles} field.
}
