% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_styles.R
\name{HTable_row_replace_style}
\alias{HTable_row_replace_style}
\title{Replace the style in a row}
\usage{
obj$row_replace_style(row = NULL, style = NULL, include_header =
  FALSE)
}
\arguments{
\item{row}{Numeric vector of which rows to target.}

\item{style}{Character vector (length 1) of style(s).}

\item{include_header}{Boolean if \code{rows} should include the header as the
1st index. E.g. if \code{FALSE}, \code{row = 1} will target the first row,
but if \code{TRUE}, \code{row = 1} will target the header.}
}
\description{
Replace the style of each <td> and, optionally, <th> tag in a
  row. Changes the \code{styles} field.
}
