% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_table_style_add}
\alias{HTable_table_style_add}
\title{Add a style to the <table> tag}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Add to the existing styles located in the \code{table_style} field.
}
