% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_table_style_replace}
\alias{HTable_table_style_replace}
\title{Replace the <table> tag style}
\usage{
obj$table_style_replace(style = NULL)
}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Overwrite the \code{table_style} field.
}
