% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_tbody_style_replace}
\alias{HTable_tbody_style_replace}
\title{Replace the <tbody> tag style}
\usage{
obj$tbody_style_replace(style = NULL)
}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Overwrite the \code{tbody_style} field.
}
