% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_tr_styles_all_clear}
\alias{HTable_tr_styles_all_clear}
\title{Clear the <tr> tag styles}
\usage{
obj$tr_styles_all_clear()

obj$tr_styles_which_clear(which = NULL)
}
\arguments{
\item{which}{Numeric vector of which <tr> tags to target.}
}
\description{
Clear any styles in the \code{tr_styles} field.

Clear any styles in the \code{tr_styles} field, but only the
  ones specified.
}
