% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_tr_styles_which_add}
\alias{HTable_tr_styles_which_add}
\title{Add styles to the <tr> tags at certain tags}
\usage{
obj$tr_styles_which_add(which = NULL, style = NULL)
}
\arguments{
\item{which}{Numeric vector of which <tr> tags to target.}

\item{style}{Character vector of style(s). It must be the same length as
\code{which}.}
}
\description{
Add to the existing styles located in the \code{tr_styles}
  field, but only to the ones specified.
}
