% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_tr_styles_which_replace}
\alias{HTable_tr_styles_which_replace}
\title{Replace the <tr> tags styles at certain tags}
\usage{
obj$tr_styles_which_replace(which = NULL, style = NULL)
}
\arguments{
\item{which}{Numeric vector of which <tr> tags to target.}

\item{style}{Character vector of style(s). It must be the same length as
\code{which}.}
}
\description{
Overwrite the \code{tr_styles} field, but only the ones
  specified.
}
