% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htable.R
\name{HTable_View}
\alias{HTable_View}
\title{View the HTable in the RStudio Viewer tab}
\usage{
obj$View()
}
\arguments{
\item{vanilla}{Should non-styled, vanilla HTML be rendered?}
}
\description{
View the HTable in the RStudio Viewer tab
}
