% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify_html}
\alias{prettify_html}
\title{Make a single string of HTML a vector with nice spacing}
\usage{
prettify_html(x, indent_char = "  ")
}
\arguments{
\item{x}{A single string of HTML}

\item{indent_char}{Character to use to indent the HTML tags.}
}
\description{
Make a single string of HTML a vector with nice spacing
}
