% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_spec.R
\name{style_spec}
\alias{style_spec}
\title{Create a style specification}
\usage{
style_spec(col_add_style_col = NULL, col_add_style_style = NULL,
  col_add_style_include_header = FALSE, col_width_col = NULL,
  col_width_width = NULL, col_bold_col = NULL,
  col_bold_include_header = FALSE, col_italic_col = NULL,
  col_italic_include_header = FALSE, col_bg_color_col = NULL,
  col_bg_color_color = NULL, col_bg_color_include_header = FALSE,
  col_color_col = NULL, col_color_color = NULL,
  col_color_include_header = FALSE, col_pct_fmt_col = NULL,
  col_pct_fmt_mult100 = TRUE, col_comma_fmt_col = NULL,
  col_dollar_fmt_col = NULL, col_color_scale_col = NULL,
  col_color_scale_color = c("#63BE7B", "#FFEB84", "#F8696B"),
  col_color_scale_exclude_rows = NULL, col_data_bar_col = NULL,
  col_data_bar_color = "lightgreen", col_data_bar_exclude_rows = NULL,
  col_centered_data_bar_col = NULL,
  col_centered_data_bar_color1 = "lightgreen",
  col_centered_data_bar_color2 = "pink",
  col_centered_data_bar_exclude_rows = NULL, row_add_style_row = NULL,
  row_add_style_style = NULL, row_bold_row = NULL,
  row_italic_row = NULL, row_bg_color_row = NULL,
  row_bg_color_color = NULL, row_color_row = NULL,
  row_color_color = NULL, row_alt_bg_color_color1 = NULL,
  row_alt_bg_color_color2 = NULL, row_alt_color_color1 = NULL,
  row_alt_color_color2 = NULL, header_add_style_style = NULL,
  header_bold_flag = FALSE, header_italic_flag = FALSE,
  header_bg_color_color = NULL, header_color_color = NULL)
}
\arguments{
\item{col_add_style_col}{Passed into \code{col_add_style}'s \code{col} argument. See \code{\link{HTable_col_add_style}}.}

\item{col_add_style_style}{Passed into \code{col_add_style}'s \code{style} argument. See \code{\link{HTable_col_add_style}}.}

\item{col_add_style_include_header}{Passed into \code{col_add_style}'s \code{include_header} argument. See \code{\link{HTable_col_add_style}}.}

\item{col_width_col}{Passed into \code{col_width}'s \code{col} argument. See \code{\link{HTable_col_width}}.}

\item{col_width_width}{Passed into \code{col_width}'s \code{width} argument. See \code{\link{HTable_col_width}}.}

\item{col_bold_col}{Passed into \code{col_bold}'s \code{col} argument. See \code{\link{HTable_col_bold}}.}

\item{col_bold_include_header}{Passed into \code{col_bold}'s \code{include_header} argument. See \code{\link{HTable_col_bold}}.}

\item{col_italic_col}{Passed into \code{col_italic}'s \code{col} argument. See \code{\link{HTable_col_italic}}.}

\item{col_italic_include_header}{Passed into \code{col_italic}'s \code{include_header} argument. See \code{\link{HTable_col_italic}}.}

\item{col_bg_color_col}{Passed into \code{col_bg_color}'s \code{col} argument. See \code{\link{HTable_col_bg_color}}.}

\item{col_bg_color_color}{Passed into \code{col_bg_color}'s \code{color} argument. See \code{\link{HTable_col_bg_color}}.}

\item{col_bg_color_include_header}{Passed into \code{col_bg_color}'s \code{include_header} argument. See \code{\link{HTable_col_bg_color}}.}

\item{col_color_col}{Passed into \code{col_color}'s \code{col} argument. See \code{\link{HTable_col_color}}.}

\item{col_color_color}{Passed into \code{col_color}'s \code{color} argument. See \code{\link{HTable_col_color}}.}

\item{col_color_include_header}{Passed into \code{col_color}'s \code{include_header} argument. See \code{\link{HTable_col_color}}.}

\item{col_pct_fmt_col}{Passed into \code{col_pct_fmt}'s \code{col} argument. See \code{\link{HTable_col_pct_fmt}}.}

\item{col_pct_fmt_mult100}{Passed into \code{col_pct_fmt}'s \code{mult100} argument. See \code{\link{HTable_col_pct_fmt}}.}

\item{col_comma_fmt_col}{Passed into \code{col_comma_fmt}'s \code{col} argument. See \code{\link{HTable_col_comma_fmt}}.}

\item{col_dollar_fmt_col}{Passed into \code{col_dollar_fmt}'s \code{col} argument. See \code{\link{HTable_col_dollar_fmt}}.}

\item{col_color_scale_col}{Passed into \code{col_color_scale}'s \code{col} argument. See \code{\link{HTable_col_color_scale}}.}

\item{col_color_scale_color}{Passed into \code{col_color_scale}'s \code{color} argument. See \code{\link{HTable_col_color_scale}}.}

\item{col_color_scale_exclude_rows}{Passed into \code{col_color_scale}'s \code{exclude_rows} argument. See \code{\link{HTable_col_color_scale}}.}

\item{col_data_bar_col}{Passed into \code{col_data_bar}'s \code{col} argument. See \code{\link{HTable_col_data_bar}}.}

\item{col_data_bar_color}{Passed into \code{col_data_bar}'s \code{color} argument. See \code{\link{HTable_col_data_bar}}.}

\item{col_data_bar_exclude_rows}{Passed into \code{col_data_bar}'s \code{exclude_rows} argument. See \code{\link{HTable_col_data_bar}}.}

\item{col_centered_data_bar_col}{Passed into \code{col_centered_data_bar}'s \code{col} argument. See \code{\link{HTable_col_centered_data_bar}}.}

\item{col_centered_data_bar_color1}{Passed into \code{col_centered_data_bar}'s \code{color1} argument. See \code{\link{HTable_col_centered_data_bar}}.}

\item{col_centered_data_bar_color2}{Passed into \code{col_centered_data_bar}'s \code{color2} argument. See \code{\link{HTable_col_centered_data_bar}}.}

\item{col_centered_data_bar_exclude_rows}{Passed into \code{col_centered_data_bar}'s \code{exclude_rows} argument. See \code{\link{HTable_col_centered_data_bar}}.}

\item{row_add_style_row}{Passed into \code{row_add_style}'s \code{row} argument. See \code{\link{HTable_row_add_style}}.}

\item{row_add_style_style}{Passed into \code{row_add_style}'s \code{style} argument. See \code{\link{HTable_row_add_style}}.}

\item{row_bold_row}{Passed into \code{row_bold}'s \code{row} argument. See \code{\link{HTable_row_bold}}.}

\item{row_italic_row}{Passed into \code{row_italic}'s \code{row} argument. See \code{\link{HTable_row_italic}}.}

\item{row_bg_color_row}{Passed into \code{row_bg_color}'s \code{row} argument. See \code{\link{HTable_row_bg_color}}.}

\item{row_bg_color_color}{Passed into \code{row_bg_color}'s \code{color} argument. See \code{\link{HTable_row_bg_color}}.}

\item{row_color_row}{Passed into \code{row_color}'s \code{row} argument. See \code{\link{HTable_row_color}}.}

\item{row_color_color}{Passed into \code{row_color}'s \code{color} argument. See \code{\link{HTable_row_color}}.}

\item{row_alt_bg_color_color1}{Passed into \code{row_alt_bg_color}'s \code{color1} argument. See \code{\link{HTable_row_alt_bg_color}}.}

\item{row_alt_bg_color_color2}{Passed into \code{row_alt_bg_color}'s \code{color2} argument. See \code{\link{HTable_row_alt_bg_color}}.}

\item{row_alt_color_color1}{Passed into \code{row_alt_color}'s \code{color1} argument. See \code{\link{HTable_row_alt_color}}.}

\item{row_alt_color_color2}{Passed into \code{row_alt_color}'s \code{color2} argument. See \code{\link{HTable_row_alt_color}}.}

\item{header_add_style_style}{Passed into \code{header_add_style}'s \code{style} argument. See \code{\link{HTable_header_add_style}}.}

\item{header_bold_flag}{Boolean if \code{header_bold} should be run. See \code{\link{HTable_header_bold}}.}

\item{header_italic_flag}{Boolean if \code{header_italic} should be run. See \code{\link{HTable_header_italic}}.}

\item{header_bg_color_color}{Passed into \code{header_bg_color}'s \code{color} argument. See \code{\link{HTable_header_bg_color}}.}

\item{header_color_color}{Passed into \code{header_color}'s \code{color} argument. See \code{\link{HTable_header_color}}.}
}
\value{
List of class "style_spec".
}
\description{
Create a reausable style specification. This makes it easier apply the same
styling to multiple htables.
}
