% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_styles.R
\name{HTable_thead_style_replace}
\alias{HTable_thead_style_replace}
\title{Replace the <thead> tag style}
\usage{
obj$thead_style_replace(style = NULL)
}
\arguments{
\item{style}{Character vector (length 1) of style(s).}
}
\description{
Overwrite the \code{thead_style} field.
}
